move OutOfOfficeNotificationUtils.js to src/misc
