+++ src/login/LoginViewController.js
@@ -37,7 +37,7 @@ import {checkApprovalStatus, getLoginErrorMessage} from "../misc/LoginUtils"
 import {getHourCycle} from "../misc/Formatter"
 import * as notificationOverlay from "../gui/base/NotificationOverlay"
 import {ButtonType} from "../gui/base/ButtonN"
-import {isNotificationCurrentlyActive, loadOutOfOfficeNotification} from "../api/main/OutOfOfficeNotificationUtils"
+import {isNotificationCurrentlyActive, loadOutOfOfficeNotification} from "../misc/OutOfOfficeNotificationUtils"
 import type {OutOfOfficeNotification} from "../api/entities/tutanota/OutOfOfficeNotification"
 import {showMoreStorageNeededOrderDialog} from "../misc/SubscriptionDialogs"
 import type {Theme} from "../gui/theme"
