+++ src/settings/EditOutOfOfficeNotificationDialog.js
@@ -11,7 +11,7 @@ import type {CheckboxAttrs} from "../gui/base/CheckboxN"
 import {CheckboxN} from "../gui/base/CheckboxN"
 import {px} from "../gui/size"
 import {ButtonType} from "../gui/base/ButtonN"
-import {getDefaultNotificationLabel} from "../api/main/OutOfOfficeNotificationUtils"
+import {getDefaultNotificationLabel} from "../misc/OutOfOfficeNotificationUtils"
 import {showBusinessFeatureRequiredDialog} from "../misc/SubscriptionDialogs"
 import type {DropDownSelectorAttrs} from "../gui/base/DropDownSelectorN"
 import {DropDownSelectorN} from "../gui/base/DropDownSelectorN"
